+++ apps/docs/src/examples/badge/badge/Badge.tsx
@@ -1,0 +1,16 @@
+import { Badge } from '@pikas-ui/badge'
+import { ExampleContainer } from '@/components/ExampleContainer'
+
+export const BadgeExample: React.FC = () => {
+  return (
+    <ExampleContainer>
+      <Badge leftIconName="bx:baguette">Bread</Badge>
+      <Badge leftIconName="bx:baguette" color="SECONDARY">
+        Bread
+      </Badge>
+      <Badge leftIconName="bx:baguette" color="TERTIARY">
+        Bread
+      </Badge>
+    </ExampleContainer>
+  )
+}