+++ apps/docs/src/examples/badge/badgeIcon/BadgeIcon.tsx
@@ -1,0 +1,12 @@
+import { BadgeIcon } from '@pikas-ui/badge'
+import { ExampleContainer } from '@/components/ExampleContainer'
+
+export const BadgeIconExample: React.FC = () => {
+  return (
+    <ExampleContainer>
+      <BadgeIcon iconName="bx:baguette" />
+      <BadgeIcon iconName="bx:baguette" color="SECONDARY" />
+      <BadgeIcon iconName="bx:baguette" color="TERTIARY" />
+    </ExampleContainer>
+  )
+}