+++ packages/badge/src/index.ts
@@ -1,0 +1,3 @@
+export * from './badge/index.js'
+export * from './badgeIcon/index.js'
+export * from './customBadge/index.js'