+++ packages/badge/src/types.ts
@@ -1,0 +1,8 @@
+export const BadgePadding = {
+  xs: true,
+  sm: true,
+  md: true,
+  lg: true,
+  xl: true,
+}
+export type BadgePaddingType = keyof typeof BadgePadding